1. change file name (ml_class.py => utils.py) 2. make TestSplit class into a function 3. merge pipeline functions into one function
